change strategy again now parallel added to Room database and to cloud backup.

+++ app/src/main/java/com/example/myapplication/network/ApiService.kt
@@ -42,7 +42,7 @@ interface ApiService {
     suspend fun addCard(@Body card: CardSchema): CardDto
 
     @GET("cards/{id}")
-    suspend fun getCard(@Path("id") cardId: Int): CardDto
+    suspend fun getCard(@Path("id") cardId: Long): CardDto
 
     @PATCH("cards")
     suspend fun updateCardBalance(@Body updateSchema: BalanceCardUpdateSchema): CardDto
